requirements.txt
tortoise-orm~=0.19.2
google-auth-httplib2
google-auth-oauthlib
google-api-python-client
python-dateutil~=2.8.2